@@ -5,9 +5,11 @@ import './tools/jest-matchers';
 
 import { sessionProm } from './tools/guarded-session';
 
-const assert = require('assert').strict;
+import { strict as assert } from 'assert';
+//const assert = require('assert').strict;
 
-const firebase = require('@firebase/testing');
+//const firebase = require('@firebase/testing');
+import * as firebase from '@firebase/testing';
 
 const FieldValue = firebase.firestore.FieldValue;
 
